Never fail to indent
A proper indentation engine should never look past the previous line.
This is because a user frequently wants custom indentation, in all
programming languages. A PureScript example:

```haskell
f1 x = case x of
  true -> true
  _ -> false

f2 x = case x of
        true -> true
        _ -> false
```

So the proper behavior is: "take previous-line indentation, then if
previous line has a modifier increase/decrease it". This is both
simple, robust, typically meets the user expectations, and is AFAIK
what all built-in major modes do.

Unfortunately we're long past that point, the purescript-mode
indentation engine parses a lot of stuff it shouldn't. And frequently
it gives a "parsing error". In my experience all those "errors" are
bugs because the PS code is perfectly valid, but disregarding if it's
a bug or an actual unfinished code in the buffer, engine should never
leave a user with no indentation whatsoever.

So what this commit does is it detects such "errors", and takes
previous indentation level. Even if it's not the correct indentation,
in practice it's only off by `purescript-indentation-left-offset`, so
a user often has much less to type compared to "no indentation" at
all.

-------

Now, it turns out the mode has another bug. As explained in the
function being added, `purescript-newline-and-indent` calls
indentation calculations on the wrong line. This isn't some special
code that only `purescript-newline-and-indent` uses, it is a code
that's being called by the <kbd>TAB</kbd> indentation as well.

As there are huge amounts of poorly documented code that shouldn't
even be there in the first place, and are sometimes using local
variables from other functions, I decided instead of trying to fix the
code it would be more productive to just detect whether we're being
called from `purescript-newline-and-indent` or not.