Implement ability to switch from .purs to .js FFI file
Emacs provides a `(ff-find-other-file nil t)` function to switch
between header and implementations. The closest PureScript has to such
idea is the FFI, where given `Foo.purs` there has to be similarly
named `Foo.js`, and unsurprisingly I keep pressing a keybind trying to
switch between them.

So add the mapping from `.purs` to `.js`.

Worth noting that this doesn't provide the reverse mapping `.js →
.purs` because js is a different major mode and we can't be making
assumptions on whether it's related to PureScript… Usually. We
probably could reassign in case the js file was open by the
`ff-find-other-file`, but let's keep it simple for now.