Implement going to the beginning of a defun
This implements support for `(beginning-of-defun)`. As a side note,
the match is done inside a single line string because I tried
implementing a regexp to make Emacs to restrain to a single line, and
it turned out to be very hard. I didn't succeed at that for like
½-hour I spent, and constraining the search artifically turned out not
just much easier, but also significantly simplified the regexp.