Remove `sexp-show` handling code
It turns out this file wasn't in the build system and was causing a
lot of warnings upon compilation. Further digging into what is that it
does revealed it is a code for handling `sexp-show` from Hackage to do
some Show-related edits in the buffer. However, it ain't clear how
it's supposed to work (sexp-show just hangs when invoked with --help)
and if it's even useful for PureScript (the file was part of the
haskell-mode codebase when the project was forked).

Original haskell-mode removed this code in 2015 here
https://github.com/haskell/haskell-mode/pull/616 with commentary:

> It requires the sexp-show package (which hasn't been updated since 2012
> and was abandoned in 2014) to be installed from Hackage to work.
> However, even if it's customization option is enabled it doesn't
> actually seem to be used/called anywhere, making it less than useful.
>
> Closes #615

Let's follow the trend.