font-lock: cover more operators than just the builtins
For example, the tuple construction `1 /\ 2` was weirdly
half-haighlighted before this commit, because it wasn't covered by the
"operators regexp" and was getting caught by some other rule. Fix that.

For inspiration, I looked at the regexp from the official Atom
purescript highlight
https://github.com/purescript-contrib/atom-language-purescript/blob/d17eee55b12c140e8a1a750cce7e5aa9b09653c2/src/purescript.coffee#L32
that one works a bit differently, it excludes chars from class. AFAIK
Emacs can't do that, so instead I just enlisted the symbols. One thing
I did differently though is including colon, because it is a valid
operator in PureScript (for List), I think Atom has a bug in their
regexp.