font-lock: highlight operators with "builtin" face
What face is the most appropriate is arguable. Emacs 29.1 has
"operator-face", but it's unavailable for older Emacs, and even for
newer versions it's just "nil", so no highlight. Kinda pointless face.

The current "variable-name" is definitely wrong though, because
operators are far from being variables. And there's a separate
"color-identifiers-mode" which uses this face to highlight similar
variables, and that would work wrong for purescript-mode.

Thus, replace the highlight to "builtin".