font-lock: remove useless rule for (,) and (->)
The (,) and (->) don't have much use in PureScript, nor in Haskell for
that matter. In Haskell (,) was a prefix notation for tuple
construction, but in PureScript this operator isn't a thing. Both
trigger "syntax error" in PureScript (and (->) doesn't work even in
Haskell).

Indeed a debatable rule. Let's just remove it.