purescript-font-lock.el: highlight docstrings unconditionally
The variable was introduced to avoid highlighting docstrings for Emacs
versions where `font-lock-doc-face` didn't exist, that is prior to 21.

It's a variable from very old times and nowadays we support Emacs
25.1+ which has font-lock-doc-face. So remove the variable and the
condition.

For the safe case I did search over Github, and the variable isn't
mentioned anywhere except the literal copies of the font-lock.el file.