Remove purescript-str module
This file is another legacy from Haskell-mode. It contains no
autoloaded or interactive (i.e. ones somebody could bind to a hotkey)
functions, and the module isn't used anywhere in the code. Some of the
functions are specific to Haskell — for example, `purescript-str-trim`
mentions some "PureScript Report", which actually refers to Haskell
report (when the mode was imported from Haskell, code was
batch-renamed from haskell to purescript). So there's no point to prefer
purescript-str-trim to plain string-trim.

Thus, let's just remove the module from the codebase.