Remove utils module
The module contains just 2 functions, none of which are used in the
project or are interactive. Of them,
`purescript-utils-parse-import-statement-at-point` contains various
Haskell-specific parsing and doesn't sound useful to a common user.
The `purescript-utils-read-directory-name` is just some odd wrapper
for normalizing a path, and unlikely to be useful to anyone outside
the project.

Thus, just remove the module.